+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## Version 0.3.0 - 2024-05-28
+
+### Features
+
+- Adding pipeine for creating dataset for disconnection-aware Chemformer model.
+- Fine-tuning of Chemformer has been updated to support Hydra configuration.
+
+### Bug-fixes
+
+- Fixed an issue in the template pipeline when sibling reaction information is not available.
+
 ## Version 0.2.0 - 2024-01-09
 
 ### Features